fixed another warning about data loss while converting

+++ include/MarketEngine.h
@@ -22,7 +22,7 @@ class MarketEngine {
         double volatility;
         double fill_probability;
     public:
-        MarketEngine(long long strategy_quote_size = 1, long long strategy_tick_offset = 1, long long strategy_max_inv = 10, long long strategy_cancel_threshold = 1, long long strategy_cooldown_between_requotes = 1, long long starting_mid_price = 10000, long long start_spread = 2, double start_vol = 1.0, double start_fill_prob = 0.3);
+        MarketEngine(int strategy_quote_size = 1, long long strategy_tick_offset = 1, long long strategy_max_inv = 10, long long strategy_cancel_threshold = 1, long long strategy_cooldown_between_requotes = 1, long long starting_mid_price = 10000, long long start_spread = 2, double start_vol = 1.0, double start_fill_prob = 0.3);
         
         void update(long long timestamp_us);
