+++ src/MarketEngine.cpp
@@ -6,7 +6,7 @@
 long long MarketEngine::env_order_id = 1000000;
 const double MarketEngine::tick_size = 0.001;
 
-MarketEngine::MarketEngine(long long strategy_quote_size, long long strategy_tick_offset, long long strategy_max_inv, long long strategy_cancel_threshold, long long strategy_cooldown_between_requotes, long long starting_mid_price, long long start_spread, double start_vol, double start_fill_prob) : orderbook(), 
+MarketEngine::MarketEngine(int strategy_quote_size, long long strategy_tick_offset, long long strategy_max_inv, long long strategy_cancel_threshold, long long strategy_cooldown_between_requotes, long long starting_mid_price, long long start_spread, double start_vol, double start_fill_prob) : orderbook(), 
                             strategy(orderbook, strategy_quote_size, strategy_tick_offset, strategy_max_inv, strategy_cancel_threshold, strategy_cooldown_between_requotes), rng(), rand_engine(rng()), market_price_ticks(starting_mid_price), spread(start_spread), volatility(start_vol), fill_probability(start_fill_prob) {}
 
 void MarketEngine::update(long long timestamp_us) {
